[IDLE-80] Retrofit2에서 실패할 경우 Error VO를 Throwing 하도록 변경
