+++ core/network/src/main/java/com/idle/network/util/onResponse.kt
@@ -1,5 +1,9 @@
 package com.idle.network.util
 
+import com.idle.domain.model.error.HttpResponseException
+import com.idle.domain.model.error.HttpResponseStatus
+import com.idle.network.model.error.ErrorResponse
+import kotlinx.serialization.json.Json
 import retrofit2.Response
 
 internal fun <T> Response<T>.onResponse(): Result<T> {
@@ -8,7 +12,14 @@ internal fun <T> Response<T>.onResponse(): Result<T> {
             return Result.success(it)
         } ?: return Result.success(Unit as T)
     } else {
-        val errorBody = errorBody()
-        throw IllegalArgumentException(errorBody.toString())
+        val errorResponse = Json.decodeFromString<ErrorResponse>(errorBody().toString())
+
+        return Result.failure(
+            HttpResponseException(
+                status = HttpResponseStatus.create(code()),
+                rawCode = errorResponse.code.toInt(),
+                msg = errorResponse.message,
+            )
+        )
     }
 }
